requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
build==1.3.0
click==8.3.1
Django==5.2.8
django-tailwind==4.4.1
djangorestframework==3.16.1
drf-yasg==1.21.7
numpy==2.3.5
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2docx==1.0.0
pdf2image==1.17.0
pdfplumber==0.11.4
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytailwindcss==0.3.0
pytesseract==0.3.13
pytz==2025.2
reportlab==4.2.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
